Add shoelace as component library.
Decided upon shoelace as a component library with well designed web components.
This should hopefully be a framework agnostic package that can speed up development and take care of accessibility on the site.

Shoelace is apparently moving towards Web Awesome. For now, this project will stick with the old library.